# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Cash Flow Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  student: "student",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
  role: UserRole;
  hasPaid: boolean;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  firstName?: string;
  lastName?: string;
  phone?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  student: "student",
  admin: "admin",
} as const;

export interface UpdateUserBody {
  role?: UpdateUserBodyRole;
  hasPaid?: boolean;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  thumbnail: string;
  price: number;
  isPublished: boolean;
  createdAt: string;
}

export interface CreateCourseBody {
  title: string;
  description: string;
  thumbnail?: string;
  price: number;
  isPublished?: boolean;
}

export type LessonVideoType =
  (typeof LessonVideoType)[keyof typeof LessonVideoType];

export const LessonVideoType = {
  youtube: "youtube",
  vimeo: "vimeo",
  upload: "upload",
  embed: "embed",
} as const;

export interface Lesson {
  id: number;
  moduleId: number;
  title: string;
  description: string;
  videoUrl: string;
  videoType: LessonVideoType;
  duration: number;
  order: number;
  isFree: boolean;
  createdAt: string;
}

export interface ModuleWithLessons {
  id: number;
  courseId: number;
  title: string;
  description: string;
  order: number;
  createdAt: string;
  lessons: Lesson[];
}

export interface CourseWithModules {
  id: number;
  title: string;
  description: string;
  thumbnail: string;
  price: number;
  isPublished: boolean;
  createdAt: string;
  modules: ModuleWithLessons[];
}

export interface Module {
  id: number;
  courseId: number;
  title: string;
  description: string;
  order: number;
  createdAt: string;
}

export interface CreateModuleBody {
  title: string;
  description: string;
  order: number;
}

export type CreateLessonBodyVideoType =
  (typeof CreateLessonBodyVideoType)[keyof typeof CreateLessonBodyVideoType];

export const CreateLessonBodyVideoType = {
  youtube: "youtube",
  vimeo: "vimeo",
  upload: "upload",
  embed: "embed",
} as const;

export interface CreateLessonBody {
  title: string;
  description: string;
  videoUrl: string;
  videoType: CreateLessonBodyVideoType;
  duration: number;
  order: number;
  isFree: boolean;
}

export interface Progress {
  id: number;
  userId: number;
  lessonId: number;
  completed: boolean;
  completedAt: string;
}

export interface MarkProgressBody {
  lessonId: number;
  completed: boolean;
}

export type CourseFileFileType =
  (typeof CourseFileFileType)[keyof typeof CourseFileFileType];

export const CourseFileFileType = {
  pdf: "pdf",
  guide: "guide",
  resource: "resource",
  video: "video",
  other: "other",
} as const;

export interface CourseFile {
  id: number;
  moduleId: number;
  name: string;
  description: string;
  fileUrl: string;
  fileType: CourseFileFileType;
  fileSize: string;
  createdAt: string;
}

export type CreateFileBodyFileType =
  (typeof CreateFileBodyFileType)[keyof typeof CreateFileBodyFileType];

export const CreateFileBodyFileType = {
  pdf: "pdf",
  guide: "guide",
  resource: "resource",
  video: "video",
  other: "other",
} as const;

export interface CreateFileBody {
  moduleId: number;
  name: string;
  description: string;
  fileUrl: string;
  fileType: CreateFileBodyFileType;
  fileSize: string;
}

export interface CaptureLeadBody {
  email: string;
  name?: string;
  phone?: string;
}

export interface DashboardSummary {
  totalCourses: number;
  totalLessons: number;
  completedLessons: number;
  progressPercent: number;
  enrolledCourses: Course[];
}

export interface GrantAccessBody {
  email: string;
  firstName?: string;
  lastName?: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  email: string;
  code: string;
  password: string;
}

export interface SetupPasswordBody {
  token: string;
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}

export interface AdminSummary {
  totalUsers: number;
  totalCourses: number;
  totalLessons: number;
  totalFiles: number;
  paidUsers: number;
  recentUsers: User[];
}

export type ListFilesParams = {
  moduleId?: number;
};