# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Cash Flow Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset code to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Code sent (always returns 200 to avoid user enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset, auto-logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/setup-password:
    post:
      operationId: setupPassword
      tags: [auth]
      summary: Set password from invitation token (new users)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupPasswordBody"
      responses:
        "200":
          description: Password set, auto-logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course with modules
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseWithModules"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /courses/{courseId}/grant-access:
    post:
      operationId: grantCourseAccess
      tags: [courses]
      summary: Grant course access to a user by email (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantAccessBody"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files (optionally by module)
      parameters:
        - name: moduleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseFile"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file record (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: File created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseFile"
  /files/{fileId}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file (admin only)
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, admin]
        hasPaid:
          type: boolean
        createdAt:
          type: string
      required: [id, email, name, role, hasPaid, createdAt]
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        isPublished:
          type: boolean
        createdAt:
          type: string
      required: [id, title, description, thumbnail, price, isPublished, createdAt]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        isPublished:
          type: boolean
      required: [title, description, price]
    CourseWithModules:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        isPublished:
          type: boolean
        createdAt:
          type: string
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ModuleWithLessons"
      required: [id, title, description, thumbnail, price, isPublished, createdAt, modules]
...
    ModuleWithLessons:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        order:
          type: integer
        createdAt:
          type: string
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required: [id, courseId, title, description, order, createdAt, lessons]
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        videoType:
          type: string
          enum: [youtube, vimeo, upload, embed]
        duration:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
        createdAt:
          type: string
      required: [id, moduleId, title, description, videoUrl, videoType, duration, order, isFree, createdAt]
...
    CourseFile:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        name:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
          enum: [pdf, guide, resource, video, other]
        fileSize:
          type: string
        createdAt:
          type: string
      required: [id, moduleId, name, description, fileUrl, fileType, fileSize, createdAt]
    CreateFileBody:
      type: object
      properties:
        moduleId:
          type: integer
        name:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
          enum: [pdf, guide, resource, video, other]
        fileSize:
          type: string
      required: [moduleId, name, description, fileUrl, fileType, fileSize]
...
    GrantAccessBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required: [email]
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required: [email]
    ResetPasswordBody:
      type: object
      properties:
        email:
          type: string
        code:
          type: string
        password:
          type: string
      required: [email, code, password]
    SetupPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required: [token, password, firstName, lastName]
...